refactor: modularize InitArgs
